+++ backend/Bookflix_Server/Bookflix_Server/Repositories/ProductoRepository.cs
@@ -68,8 +68,8 @@ namespace Bookflix_Server.Repositories
             string autor = null,
             string genero = null,
             string isbn = null,
-            decimal? precioMin = null,
-            decimal? precioMax = null,
+            double? precioMin = null,
+            double? precioMax = null,
             string ordenPor = null,
             bool ascendente = true)
         {
